Refactored SampleConverter.
